Server library now compiles on VS2010 64bit without compiler warnings (except in the hid submodule).

+++ vrpn_Magellan.C
@@ -48,7 +48,7 @@
 // This routine writes out the characters slowly, so as not to
 // overburden the poor Magellan, which seems to choke when a
 // bunch of characters are all sent at once.
-static	int	vrpn_write_slowly(int fd, unsigned char *buffer, int len, int MsecWait)
+static	int	vrpn_write_slowly(int fd, unsigned char *buffer, size_t len, int MsecWait)
 {	int	i;
 
 	for (i = 0; i < len; i++) {
@@ -57,7 +57,7 @@ static	int	vrpn_write_slowly(int fd, unsigned char *buffer, int len, int MsecWai
 			return -1;
 		}
 	}
-	return len;
+	return static_cast<int>(len);
 }
